+++ src/main/kotlin/blackjack/domain/card/Pattern.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.card
+
+enum class Pattern {
+    CLOVER, DIAMOND, HEART, SPADE,
+}
